Refactoring bootstrap.kt to domain specific classes and files.
Also, unnecessary files to the project and adjusting to kotling conventions.
