+++ backend/src/Customer.kt
@@ -1,0 +1,5 @@
+package challenge
+
+class Address
+
+class Customer
