+++ backend/src/Payment.kt
@@ -1,0 +1,19 @@
+package challenge
+
+import java.util.*
+
+data class Payment(val order: Order, val paymentMethod: PaymentMethod) {
+    val paidAt = Date()
+    val authorizationNumber = paidAt.time
+    val amount = order.totalAmount
+    val invoice = Invoice(order)
+}
+
+data class CreditCard(val number: String) : PaymentMethod
+
+interface PaymentMethod
+
+data class Invoice(val order: Order) {
+    val billingAddress: Address = order.address
+    val shippingAddress: Address = order.address
+}
