+++ backend/src/Product.kt
@@ -1,0 +1,10 @@
+package challenge
+
+data class Product(val name: String, val type: ProductType, val price: Double)
+
+enum class ProductType {
+    PHYSICAL,
+    BOOK,
+    DIGITAL,
+    MEMBERSHIP
+}
